nuevo-backend/pyproject.toml
[project]
name = "mcp-db"
version = "0.1.0"
description = "MCP Server for Database and Gmail Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.3",
    "psycopg2>=2.9.10",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.104.0",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
]
